requirements.txt
alembic==1.14.1
aniso8601==10.0.1
anyio==4.5.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.8.2
cached-property==1.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.8
decorator==5.2.1
distlib==0.3.9
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
executing==2.2.0
Faker==35.2.2
fastapi==0.79.0
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
gunicorn==21.2.0
h11==0.16.0
idna==2.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
lxml==4.9.4
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packages==0.1.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pydantic==1.10.22
Pygments==2.19.1
PyJWT==2.9.0
pymongo==3.13.0
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
redis==3.5.3
requests==2.31.0
requests-cache==0.5.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.19.1
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==1.25.11
uvicorn==0.18.2
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.0.6
whitenoise==6.7.0
zipp==3.20.2
